pyproject.toml
[project]
name = "happiness-in-space"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "pillow>=11.2.1",
    "requests>=2.32.3",
]
